Handle redirects in git clone commands
Add the repoAssignment context handler to the route that handles
the git http protocol so that when a user clones a repo that has
been moved or renamed, they are redirected to the new location.

This requires that the query string be included in the redirect
as well, so that is added.

Signed-off-by: James E. Blair <jeblair@redhat.com>